moved the methods from useEffect into directory InitialData
