+++ src/InitialData/initialData.ts
@@ -1,0 +1,12 @@
+import { IInitialCircle, IInitialGameStatic } from "../Types/Types";
+
+export const InitialCircle: IInitialCircle = {
+  extraClass: "",
+  clicked: false,
+};
+
+export const InitialGameStatic: IInitialGameStatic = {
+  number: "",
+  fillLine: false,
+  editable: true,
+};
